+++ CHANGELOG.md
@@ -6,7 +6,10 @@ All notable changes to **Enumeration** will be documented in this file. The proj
 
 ## [Unreleased]
 
-### Added
+## [1.4.0] - 2026-04-10
+
+\### Added
+
 - Option to generate JsonConverter via `GenerateJsonConverter` property on the attribute.
 - `ISpanFormattable` added
 - `DebuggerDisplay` attribute added to generated types for better debugging experience.
@@ -55,7 +58,8 @@ All notable changes to **Enumeration** will be documented in this file. The proj
 
 - Initial release of `Enumeration` source code generator.
 
-[unreleased]: https://github.com/linkdotnet/Enumeration/compare/1.3.1...HEAD
+[unreleased]: https://github.com/linkdotnet/Enumeration/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/linkdotnet/Enumeration/compare/1.3.1...1.4.0
 [1.3.1]: https://github.com/linkdotnet/Enumeration/compare/1.3.0...1.3.1
 [1.3.0]: https://github.com/linkdotnet/Enumeration/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/linkdotnet/Enumeration/compare/1.1.0...1.2.0